Remove mnist_dataset_function.py and mnist_resnet50.py.
