+++ examples/torch/MNIST/config.yaml
@@ -7,15 +7,63 @@ trainer:
       save_top_k: -1
 
 task:
-  _target_: mnist_resnet50.MNISTResNet50
+  _target_: torchvision_task.Resnet50Task
+  model:
+    _target_: torchvision.models.resnet50
+    weights:
+      _target_: hydra.utils.get_object
+      path: torchvision.models.ResNet50_Weights.DEFAULT
+  optimizer_config:
+    _target_: aiaccel.torch.lightning.OptimizerConfig
+    optimizer_generator:
+      _partial_: True
+      _target_: torch.optim.AdamW
+      lr: 1.e-4
+  num_classes: 10
 
 datamodule:
   _target_: aiaccel.torch.lightning.datamodules.single_datamodule.SingleDataModule
   train_dataset_fn:
     _partial_: true
-    _target_: mnist_dataset_function.train_dataset
+    _target_: torchvision.datasets.MNIST
+    root: "./dataset"
+    train: True
+    download: True
+    transform:
+      _target_: torchvision.transforms.Compose
+      transforms:
+        - _target_: torchvision.transforms.Resize
+          size:
+            - 224
+            - 224
+        - _target_: torchvision.transforms.Grayscale
+          num_output_channels: 3
+        - _target_: torchvision.transforms.ToTensor
+        - _target_: torchvision.transforms.Normalize
+          mean:
+            - 0.5
+          std:
+            - 0.5
   val_dataset_fn:
     _partial_: true
-    _target_: mnist_dataset_function.val_dataset
+    _target_: torchvision.datasets.MNIST
+    root: "./dataset"
+    train: False
+    download: True
+    transform:
+      _target_: torchvision.transforms.Compose
+      transforms:
+        - _target_: torchvision.transforms.Resize
+          size:
+            - 224
+            - 224
+        - _target_: torchvision.transforms.Grayscale
+          num_output_channels: 3
+        - _target_: torchvision.transforms.ToTensor
+        - _target_: torchvision.transforms.Normalize
+          mean:
+            - 0.5
+          std:
+            - 0.5
   batch_size: 2
   wrap_scatter_dataset: False
